The Fallen of World War II - Data Visualization
This project is quite near and dear to me. Learning about the reality of the horrors of war and conflict for soldiers and civilians forces you to think about what you are faithful. The creator of this project, Neil Halloran, utilizes data in such a powerful way. While the video plays, you can pause and interact with the graphs (provided you have a good computer). This element of human computer interaction reminds me of the graph we looked at that displayed the Native American land.

The film was written, directed, coded, and narrated by Neil Halloran. It also mentions data volunteers. Due to the website being one page, it is realistic that Neil created it himself.

Upon inspecting the source code, I see HTML as well as javascript. There are also a lot of the <div> commands like we went over in class. Images of my findings are located in the images-source folder. I have never used Chrome Inspect before so it was cool to see the yoke of a website and have somewhat of a clue on what everything means.